Chipping away at ability to completely delete gaming groups
