+++ Source/BusinessLogic/BusinessLogic.csproj
@@ -327,8 +327,8 @@
     <Compile Include="Models\Players\HomePagePlayerSummary.cs" />
     <Compile Include="Logic\Players\IHomePagePlayerSummaryRetriever.cs" />
     <Compile Include="Logic\Players\HomePagePlayerSummaryRetriever.cs" />
-    <Compile Include="Logic\GamingGroups\GamingGroupDeleter.cs" />
-    <Compile Include="Logic\GamingGroups\IGamingGroupDeleter.cs" />
+    <Compile Include="Logic\GamingGroups\DeleteGamingGroupComponent.cs" />
+    <Compile Include="Logic\GamingGroups\IDeleteGamingGroupComponent.cs" />
     <Compile Include="Migrations\201710040027508_GamingGroup_Active_Flag.cs" />
     <Compile Include="Migrations\201710040027508_GamingGroup_Active_Flag.Designer.cs">
       <DependentUpon>201710040027508_GamingGroup_Active_Flag.cs</DependentUpon>
