+++ Source/BusinessLogic/Logic/GamingGroups/IDeleteGamingGroupComponent.cs
@@ -1,0 +1,11 @@
+﻿using BusinessLogic.DataAccess;
+using BusinessLogic.Models.User;
+
+namespace BusinessLogic.Logic.GamingGroups
+{
+    public interface IDeleteGamingGroupComponent
+    {
+        void Execute(int gamingGroupId, ApplicationUser currentUser, IDataContext dataContext);
+        void Execute(int gamingGroupId, ApplicationUser currentUser);
+    }
+}
